requirements.txt
Flask
python-dotenv
mysql-connector-python
blinker
click
itsdangerous
jinja2
markupsafe
werkzeug
